Add publisher for basic tag detections (id and pixel coordinates)
Basic detections mean that we only report the tag id and the pixel
coordinates of the center of the tag. This will allow us to develop
codet that does not rely on the relative pose of the tag, but only on
the position of the tag in the camer image.